m111: Fix compilation errors
