+++ skia-org/src/drivers/gl.rs
@@ -1,5 +1,5 @@
 use crate::{artifact, drivers::DrawingDriver, Driver};
-use skia_safe::{gpu, Budgeted, Canvas, ImageInfo, Surface};
+use skia_safe::{gpu, Canvas, ImageInfo, Surface};
 use std::path::Path;
 
 pub struct OpenGl {
@@ -25,7 +25,7 @@ impl DrawingDriver for OpenGl {
         let image_info = ImageInfo::new_n32_premul((width * 2, height * 2), None);
         let mut surface = Surface::new_render_target(
             &mut self.context,
-            Budgeted::Yes,
+            gpu::Budgeted::Yes,
             &image_info,
             None,
             gpu::SurfaceOrigin::BottomLeft,
