+++ skia-org/src/drivers/vulkan.rs
@@ -3,7 +3,7 @@ use ash::{
     vk::{self, Handle},
     Entry, Instance,
 };
-use skia_safe::{gpu, Budgeted, Canvas, ImageInfo, Surface};
+use skia_safe::{gpu, Canvas, ImageInfo, Surface};
 use std::{convert::TryInto, ffi::CString, os::raw, path::Path, ptr};
 
 #[allow(dead_code)]
@@ -61,7 +61,7 @@ impl DrawingDriver for Vulkan {
         let image_info = ImageInfo::new_n32_premul((width * 2, height * 2), None);
         let mut surface = Surface::new_render_target(
             &mut self.context,
-            Budgeted::Yes,
+            gpu::Budgeted::Yes,
             &image_info,
             None,
             gpu::SurfaceOrigin::TopLeft,
