+++ skia-safe/src/core/types.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use skia_bindings::{SkBudgeted, SkFourByteTag};
+use skia_bindings::SkFourByteTag;
 use std::ops::Deref;
 
 #[allow(non_camel_case_types)]
@@ -81,22 +81,8 @@ pub type Unichar = skia_bindings::SkUnichar;
 // pub(crate) type MSec = skia_bindings::SkMSec;
 // pub(crate) const MSEC_MAX: u32 = std::i32::MAX as u32;
 
-#[repr(C)]
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub struct Budgeted(bool);
-
-native_transmutable!(SkBudgeted, Budgeted, budgeted_layout);
-
-#[allow(non_upper_case_globals)]
-impl Budgeted {
-    #[deprecated(since = "0.29.0", note = "use No")]
-    pub const NO: Budgeted = Budgeted(false);
-    #[deprecated(since = "0.29.0", note = "use Yes")]
-    pub const YES: Budgeted = Budgeted(true);
-
-    // we want this look like enum case names.
-    pub const No: Budgeted = Budgeted(false);
-    pub const Yes: Budgeted = Budgeted(true);
-}
+#[cfg(feature = "gpu")]
+#[deprecated(since = "0.0.0", note = "Use gpu::Budgeted")]
+pub type Budgeted = crate::gpu::Budgeted;
 
 // TODO: BackingFit
