+++ skia-safe/src/gpu/types.rs
@@ -7,6 +7,24 @@ variant_name!(BackendAPI::Dawn, gr_backend_api_naming);
 pub use skia_bindings::skgpu_BackendApi as BackendApi;
 variant_name!(BackendApi::Metal, skgpu_backend_api_naming);
 
+#[repr(C)]
+#[derive(Copy, Clone, PartialEq, Eq, Debug)]
+pub struct Budgeted(bool);
+
+native_transmutable!(sb::skgpu_Budgeted, Budgeted, budgeted_layout);
+
+#[allow(non_upper_case_globals)]
+impl Budgeted {
+    #[deprecated(since = "0.29.0", note = "use No")]
+    pub const NO: Budgeted = Budgeted(false);
+    #[deprecated(since = "0.29.0", note = "use Yes")]
+    pub const YES: Budgeted = Budgeted(true);
+
+    // we want this look like enum case names.
+    pub const No: Budgeted = Budgeted(false);
+    pub const Yes: Budgeted = Budgeted(true);
+}
+
 // TODO: this should be a newtype(bool) I guess with implementations
 //       of From<bool> and Deref?
 pub use skia_bindings::GrMipmapped as Mipmapped;
